Add files via upload
I've included the files I've been using for the Menu so far, The only files i couldn't upload due to size was, LA_Crime_Data_2023_to_Present_data which is crucial for the initial model (This model used here is outdated but the menu program should still be able to use the newer rendition of the Model when it is completed so for now think of it as a test program for the real deal) So please download The data set separately from the Canvas page and run it with the Data_Cleaning_Notebook to get Cleaned_Dataset which is used for the menu to 
1: Load 
2: Load Model 
3: "Train" the model 
The program to use is named CMPS_3500_FINALTE.py TE for test as it still has some working need done, The way i've been running is that i have an environment with Python 3.11.0 to make use of all the requirements listed on the requirements.txt file to run this just use 
pip install -r requirements.txt within the environment you make so that the program can be ran
Let me know if you want me to help with the environment making!
All that is missing is the New Model functionality and anything else we think may be missing, Please read the documentation. Any question that may come up please as in the Disc Chat ill attempt to answer to the best of my ability! Let's get this done yall!